refactor(test): separate RequestLogITest
